最后一次修改 将网址一拆为十，每个脚本修改后的代码没有合并，后续使用可以将十个脚本针对异常修改的代码进行合并

+++ python4_16/girlsV4/girlsV4_2.py
@@ -11,34 +11,33 @@ class StaticPicture:
   num2=0
   # 自定义文件夹名称(  网页地址，    网页内容 ,文件夹名称，  文件夹正则, 路径文件正则)
   def dir_name (self,url:str,request,dir_name:str,ze1:str,lujing:str):
-    request.encoding = request.apparent_encoding
+
     html=request.text
     list2=re.findall('http://www\.cgtpw\.com/(.*?)/.*?\.html',url)[-1]
     if len(url) ==0:
       return lujing+dir_name
     else:
       try:
-        return lujing+list2+'\\'+re.findall(ze1,html)[-1]
+        return lujing+list2+'\\'+re.sub('([\\/:"*?<>|\x00-\x1F]+)','',re.findall(ze1,html)[-1])#解决文件夹名称出现异常字符问题
       except:
-        return ''
+        return 'D:\\mn\\'+list2+'\\test'
   # ze:符合条件的图片格式 <p align="center"><img src="(.*?)" alt="" /></p>
   #time:单次下载需要等待的时间单位为秒
   def downLoad(self,url:str,request,dir_name:str,ze:str):
-    request.encoding = request.apparent_encoding
+
+
     HTML=request.text
     list2=re.findall('http://www\.cgtpw\.com/(.*?)/.*?\.html',url)[-1]
     urls=[]
     # 创建文件夹
-    if os.path.exists(dir_name):
-      if len(os.listdir(dir_name))==0:
-        dir_name=dir_name
-      else:
-        return ''
     if not os.path.exists(dir_name):
       try:
         os.mkdir(dir_name)
       except FileNotFoundError:
-        os.makedirs(dir_name)
+        try:
+          os.makedirs(dir_name)
+        except:
+          dir_name='D:\\mn\\'+list2+'\\test'
       except:
         dir_name='D:\\mn\\'+list2+'\\test'
     # 正则匹配到所有的图片地址
@@ -57,38 +56,50 @@ class StaticPicture:
     for tmp in tmp9:
       urls.append(tmp[tmp5])
     # 遍历下载
-    for url in urls:
-      print(url)
-      file_name=url.split('/')[-1]
+    for url1 in urls:
+      file_name=url1.split('/')[-1]
       try:
-        file_name=re.findall('(.*?)" .*?',file_name)[-1]
+        file_name=re.findall('(.*?\.jpg)',file_name)[-1]
       except:
         return ""
-      print(file_name)
-      with open(dir_name+'/'+file_name,'wb') as f:
-        f.write(request.content)
+      str2=dir_name+'/'+file_name
+      if  os.path.exists(str2):
+        return ""
+      else:
+        try:
+          request1= requests.get(url1)
+        except:
+          return ''
+        with open(dir_name+'/'+file_name,'wb') as f:
+          f.write(request1.content)
+          print(url1)
 
 
 
 t=StaticPicture()
-t.num1=272833
 sum_list5=[]
 f = open("foo_sum_2.txt") # 返回一个文件对象
 fo=open('foo_sum_2_download.txt', "a")
+fo1=open('foo_sum_2_blank.txt','a')
 line = f.readline() # 调用文件的 readline()方法
 while line:
   fo.write(line)#写入其他文件
   line=line.replace('\n', '')#替换掉换行符
   #通过URL获取到HTML内容
-  request=requests.get(line,headers=headers,timeout=10)
+  request=requests.get(line,headers=headers)
   request.encoding = request.apparent_encoding
   #下载图片
-  dir_name=t.dir_name(line,request,"",'<h1>(.*?)</h1>','D:\\mn\\')
+  dir_name=t.dir_name(line,request,"",'<h1>([\\/:"*?<>|\x00-\x1F]+)</h1>','D:\\mn\\')
+  print("文件名成功为"+dir_name)
   t.num2=t.num2+1
-  print("下载总进度"+str(t.num2)+'/'+str(t.num1))
+  print("下载总进度"+str(t.num2))
   t.downLoad(line,request,dir_name,'<img src="(.*?)" alt=".*?"( class=".*?"| /)>')
+  print(dir_name)
+  if os.path.getsize(dir_name)==0:
+    fo1.write(line+'\n')
   #继续读取下一行
   line = f.readline()
+fo1.close()
 fo.close()
 f.close()
